KUSH GALAXY 6.66 Introduction / README.txt
A small synopsis on what the actual point of this beginner project is. 
Also includes instructions and controls for the game. 

Thanks for checking this out ! 

I've had a lot of fun in this one ! 
Biggest lesson I've learned here is that finding the solution of an error is the most satisfying thing in the whole process.

Thanks for checking this out !